requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
google-generativeai==0.2.0
pandas==2.1.3
